app/requirements.txt
pyspark
kafka-python
cassandra-driver
python-dotenv
